requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
decouple==0.0.7
Django==4.2.3
django-crispy-forms==2.0
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
